requirements.txt
# Add your requirements here like:
numpy
jpype1
urllib
